aws: move flb_read_file to utils
