+++ src/aws/flb_aws_credentials.c
@@ -26,8 +26,6 @@
 #include <jsmn/jsmn.h>
 #include <stdlib.h>
 #include <time.h>
-#include <sys/types.h>
-#include <sys/stat.h>
 
 #define TEN_MINUTES    600
 #define TWELVE_HOURS   43200
@@ -533,53 +531,6 @@ time_t flb_aws_cred_expiration(const char *timestamp)
                   timestamp);
      }
      return expiration;
-}
-
-int flb_read_file(const char *path, char **out_buf, size_t *out_size)
-{
-    int ret;
-    long bytes;
-    char *buf = NULL;
-    FILE *fp = NULL;
-    struct stat st;
-    int fd;
-
-    fp = fopen(path, "r");
-    if (!fp) {
-        return -1;
-    }
-
-    fd = fileno(fp);
-    ret = fstat(fd, &st);
-    if (ret == -1) {
-        flb_errno();
-        fclose(fp);
-        return -1;
-    }
-
-    buf = flb_malloc(st.st_size + sizeof(char));
-    if (!buf) {
-        flb_errno();
-        fclose(fp);
-        return -1;
-    }
-
-    bytes = fread(buf, st.st_size, 1, fp);
-    if (bytes != 1) {
-        flb_errno();
-        flb_free(buf);
-        fclose(fp);
-        return -1;
-    }
-
-    /* fread does not add null byte */
-    buf[st.st_size] = '\0';
-
-    fclose(fp);
-    *out_buf = buf;
-    *out_size = st.st_size;
-
-    return 0;
 }
 
 /*
